Extracted parsing logic out of semver data class. See #14
