+++ bumpit/core/versions/semver/__init__.py
@@ -1,4 +1,4 @@
-from bumpit.core.versions.semver.semver import SemVer
+from bumpit.core.versions.semver.parsers import parse
 from bumpit.core.versions.semver.transformers import (
     SemverIncrementingTransformer,
     SemverStaticTransformer,
@@ -8,7 +8,7 @@ SEMVER_STRATEGY = "semver"
 
 
 def next_semantic_version(current_version, part, force_value, _=None):
-    version = SemVer.parse(current_version)
+    version = parse(current_version)
     if force_value is None:
         transform = SemverIncrementingTransformer()
         return str(transform(part, version))
