+++ bumpit/core/versions/semver/semver.py
@@ -1,4 +1,3 @@
-import re
 from dataclasses import dataclass
 
 
@@ -9,21 +8,6 @@ class SemVer:
     patch: int
     pre_release: str
     build_metadata: str
-
-    @staticmethod
-    def parse(version):
-        semver_pattern = "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\+([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$"  # noqa
-        match = re.search(semver_pattern, version)
-        if not match:
-            raise ValueError(f"Invalid semantic version '{version}'.")
-
-        return SemVer(
-            major=int(match.group(1)),
-            minor=int(match.group(2)),
-            patch=int(match.group(3)),
-            pre_release=match.group(4) or "",
-            build_metadata=match.group(5) or "",
-        )
 
     def __str__(self):
         as_str = f"{self.major}.{self.minor}.{self.patch}"
