+++ bumpit/core/versions/semver/transformers.py
@@ -1,4 +1,4 @@
-from bumpit.core.versions.semver.semver import SemVer
+from bumpit.core.versions.semver.parsers import parse
 
 NUMERICAL_PARTS = ["major", "minor", "patch"]
 NON_NUMERICAL_PARTS = ["pre_release", "build_metadata"]
@@ -32,7 +32,7 @@ class SemverStaticTransformer:
         if getattr(version, part) >= value:
             raise ValueError(f"Can only increase {part} part.")
 
-        new_version = SemVer.parse("0.0.0")
+        new_version = parse("0.0.0")
 
         target_part_index = NUMERICAL_PARTS.index(part)
         for current_index, current_part in enumerate(NUMERICAL_PARTS):
@@ -49,7 +49,7 @@ class SemverStaticTransformer:
 
     @staticmethod
     def _transform_non_numerical_part(part, version, static):
-        new_version = SemVer.parse(str(version))
+        new_version = parse(str(version))
         if getattr(version, part) == static:
             raise ValueError("There is no version change.")
